Fix bug with adjusting cells font size
Initially, after changing the grid size, the cells font size didn't get
adjusted correctly, because it would take the measurements of the cells
before the DOM would get updated, which meant that the font was
calculated based on the previous grid size.

The solution to that was to make adjustCellsFontSize function to use the
grid height and divide it by the grid size to get a hight of one cell.

That revealed another problem, which was that when adding
adjustCellsFontSize as event listeners to the window, it created
closure, which meant that on window resize it would adjust the cells
font size relying on the grid size that was provided the first time
(i.e. on component mount), and didn't not use the current grid size
value.

To fix that, instead of adding event listeners just once on the
component mount, add and remove them through useEffect with size as it's
dependency. That way, the function that would handle the event in it's
closure would have access to the correct grid size.